+++ src/com/company/Main.java
@@ -1,11 +1,17 @@
 package com.company;
+
+import java.util.HashSet;
 import java.util.Scanner;
 
 public class Main {
+
     public static void main(String[] args) {
-        Scanner in = new Scanner(System.in);
-        outputMenu(); // Вывод меню
+        First first = new First();
+        Second second = new Second();
+        //Third third = new Third();
         int command = -1;
+        Scanner in = new Scanner(System.in);
+        outputMenu();
         while(command != 0){
             System.out.print("Введите команду: ");
             command = in.nextInt();
@@ -14,46 +20,44 @@ public class Main {
                     outputMenu();
                     break;
                 case 2:
-                    try{
-                        //something
-                    }catch (Exception e) {
-                        System.out.println(e);
-                    }
+                    System.out.print("Введите текст: ");
+                    first.setText(in.nextLine());
                     break;
                 case 3:
-                    try{
-                        //something
-                    }catch (Exception e) {
-                        System.out.println(e);
-                    }
+                    first.vowelWords();
                     break;
                 case 4:
-                    try{
-                        //something
-                    }catch (Exception e) {
-                        System.out.println(e);
-                    }
+                    System.out.print("Введите слово: ");
+                    first.longerThan(in.nextLine());
                     break;
                 case 5:
-                    try{
-                        //something
-                    }catch (Exception e) {
-                        System.out.println(e);
-                    }
+                    first.upperPrePri();
                     break;
                 case 6:
-                    try{
-                        //something
-                    }catch (Exception e) {
-                        System.out.println(e);
-                    }
+                    System.out.print("Введите диапазон: ");
+                    first.numberRange(in.nextInt(),in.nextInt());
                     break;
                 case 7:
-                    try{
-                        //something
-                    }catch (Exception e) {
-                        System.out.println(e);
-                    }
+                    System.out.print("Введите текст: ");
+                    second.setText(in.nextLine());
+                    break;
+                case 8:
+                    second.removePrePri();
+                    break;
+                case 9:
+                    second.divSev();
+                    break;
+                case 10:
+                    second.addZeros();
+                    break;
+                case 11:
+                    //some
+                    break;
+                case 12:
+                    //some
+                    break;
+                case 13:
+                    //some
                     break;
                 case 0:
                     System.out.println("Завершение программы...");
@@ -65,14 +69,21 @@ public class Main {
         }
     }
 
+
     public static void outputMenu(){
-        System.out.println("1 ");
-        System.out.println("2 ");
-        System.out.println("3 ");
-        System.out.println("4 ");
-        System.out.println("5 ");
-        System.out.println("6 ");
-        System.out.println("7 ");
-        System.out.println("0 - Выход.");
+        System.out.println("1 - Вывести меню повторно");
+        System.out.println("2 - Ввести предложения(Задание №1)");
+        System.out.println("3 - Сформировать предложение из слов, которые начинаются и заканчиваются гласной буквой.");
+        System.out.println("4 - Сколько в сформированном предложении слов, больше заданного слова.");
+        System.out.println("5 - Все слова исходного текста, которые начинаются с приставок пре и при записать прописными буквами.");
+        System.out.println("6 - Определить, сколько чисел текста принадлежат заданному диапазону.");
+        System.out.println("7 - Ввести предложения(Задание №2)");
+        System.out.println("8 - Удалить из текста все слова, которые начинаются с приставок пре и при.");
+        System.out.println("9 - Целые числа, которые не кратны 7, заменить результатом этого деления.");
+        System.out.println("10 - Добавить строку из семи нулей, перед каждым словом не числом.");
+        System.out.println("11 ");
+        System.out.println("12 ");
+        System.out.println("13 ");
+        System.out.println("0 - Выход");
     }
 }
